backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12
djangorestframework-simplejwt>=4.7
psycopg2-binary>=2.9
django-cors-headers>=3.7
redis>=3.5
gunicorn>=20.1
